Make logout reusable
